backend/dsl/enh/Rider side MerchantServiceConfig to dsl
