+++ Backend/app/rider-platform/rider-app/Main/src-read-only/Storage/Queries/MerchantServiceConfig.hs
@@ -1,0 +1,12 @@
+{-# OPTIONS_GHC -Wno-dodgy-exports #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
+{-# OPTIONS_GHC -Wno-unused-imports #-}
+
+module Storage.Queries.MerchantServiceConfig (module Storage.Queries.MerchantServiceConfig, module ReExport) where
+
+import Kernel.Beam.Functions
+import Kernel.External.Encryption
+import Kernel.Prelude
+import Kernel.Types.Error
+import Kernel.Utils.Common (CacheFlow, EsqDBFlow, MonadFlow, fromMaybeM, getCurrentTime)
+import Storage.Queries.MerchantServiceConfigExtra as ReExport
