+++ Backend/app/rider-platform/rider-app/Main/src-read-only/Storage/Queries/OrphanInstances/PlaceBasedServiceConfig.hs
@@ -1,4 +1,3 @@
-{-# OPTIONS_GHC -Wno-deprecations #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 {-# OPTIONS_GHC -Wno-unused-imports #-}
 
@@ -12,11 +11,11 @@ import Kernel.Types.Error
 import qualified Kernel.Types.Id
 import Kernel.Utils.Common (CacheFlow, EsqDBFlow, MonadFlow, fromMaybeM, getCurrentTime)
 import qualified Storage.Beam.PlaceBasedServiceConfig as Beam
-import qualified Storage.Queries.Merchant.MerchantServiceConfig
+import qualified Storage.Queries.Transformers.MerchantServiceConfig
 
 instance FromTType' Beam.PlaceBasedServiceConfig Domain.Types.PlaceBasedServiceConfig.PlaceBasedServiceConfig where
   fromTType' (Beam.PlaceBasedServiceConfigT {..}) = do
-    serviceConfig' <- Storage.Queries.Merchant.MerchantServiceConfig.getServiceConfigFromDomain serviceName configValue
+    serviceConfig' <- Storage.Queries.Transformers.MerchantServiceConfig.getServiceConfigFromDomain serviceName configValue
     pure $
       Just
         Domain.Types.PlaceBasedServiceConfig.PlaceBasedServiceConfig
@@ -34,8 +33,8 @@ instance ToTType' Beam.PlaceBasedServiceConfig Domain.Types.PlaceBasedServiceCon
       { Beam.merchantId = Kernel.Types.Id.getId merchantId,
         Beam.merchantOperatingCityId = Kernel.Types.Id.getId merchantOperatingCityId,
         Beam.placeId = Kernel.Types.Id.getId placeId,
-        Beam.configValue = snd $ Storage.Queries.Merchant.MerchantServiceConfig.getServiceNameConfigJson serviceConfig,
-        Beam.serviceName = fst $ Storage.Queries.Merchant.MerchantServiceConfig.getServiceNameConfigJson serviceConfig,
+        Beam.configValue = snd $ Storage.Queries.Transformers.MerchantServiceConfig.getServiceNameConfigJson serviceConfig,
+        Beam.serviceName = fst $ Storage.Queries.Transformers.MerchantServiceConfig.getServiceNameConfigJson serviceConfig,
         Beam.createdAt = createdAt,
         Beam.updatedAt = updatedAt
       }
