+++ Backend/app/rider-platform/rider-app/Main/src/Domain/Action/Dashboard/Merchant.hs
@@ -27,10 +27,10 @@ import qualified "dashboard-helper-api" Dashboard.RiderPlatform.Merchant as Comm
 import qualified Domain.Types.Exophone as DExophone
 import qualified Domain.Types.Geometry as DGEO
 import qualified Domain.Types.Merchant as DM
-import qualified Domain.Types.Merchant.MerchantServiceConfig as DMSC
 import qualified Domain.Types.MerchantMessage as DMM
 import qualified Domain.Types.MerchantOperatingCity as DMOC
 import qualified Domain.Types.MerchantPaymentMethod as DMPM
+import qualified Domain.Types.MerchantServiceConfig as DMSC
 import qualified Domain.Types.MerchantServiceUsageConfig as DMSUC
 import qualified Domain.Types.RiderConfig as DRC
 import Environment
@@ -131,6 +131,23 @@ mkServiceUsageConfigRes DMSUC.MerchantServiceUsageConfig {..} =
     }
 
 ---------------------------------------------------------------------
+buildMerchantServiceConfig ::
+  MonadTime m =>
+  Id DM.Merchant ->
+  Id DMOC.MerchantOperatingCity ->
+  DMSC.ServiceConfig ->
+  m DMSC.MerchantServiceConfig
+buildMerchantServiceConfig merchantId merchantOperatingCityId serviceConfig = do
+  now <- getCurrentTime
+  pure
+    DMSC.MerchantServiceConfig
+      { merchantId,
+        serviceConfig,
+        merchantOperatingCityId,
+        updatedAt = now,
+        createdAt = now
+      }
+
 mapsServiceConfigUpdate ::
   ShortId DM.Merchant ->
   Context.City ->
@@ -143,7 +160,7 @@ mapsServiceConfigUpdate merchantShortId city req = do
   merchantOperatingCity <-
     CQMOC.findByMerchantShortIdAndCity merchantShortId city
       >>= fromMaybeM (MerchantOperatingCityNotFound ("merchantShortId: " <> merchantShortId.getShortId <> " ,city: " <> show city))
-  merchantServiceConfig <- DMSC.buildMerchantServiceConfig merchant.id merchantOperatingCity.id serviceConfig
+  merchantServiceConfig <- buildMerchantServiceConfig merchant.id merchantOperatingCity.id serviceConfig
   _ <- CQMSC.upsertMerchantServiceConfig merchantServiceConfig
   CQMSC.clearCache merchant.id merchantOperatingCity.id serviceName
   logTagInfo "dashboard -> mapsServiceConfigUpdate : " (show merchant.id)
@@ -162,7 +179,7 @@ smsServiceConfigUpdate merchantShortId city req = do
       >>= fromMaybeM (MerchantOperatingCityNotFound ("merchantShortId: " <> merchantShortId.getShortId <> " ,city: " <> show city))
   let serviceName = DMSC.SmsService $ Common.getSmsServiceFromReq req
   serviceConfig <- DMSC.SmsServiceConfig <$> Common.buildSmsServiceConfig req
-  merchantServiceConfig <- DMSC.buildMerchantServiceConfig merchant.id merchantOperatingCity.id serviceConfig
+  merchantServiceConfig <- buildMerchantServiceConfig merchant.id merchantOperatingCity.id serviceConfig
   _ <- CQMSC.upsertMerchantServiceConfig merchantServiceConfig
   CQMSC.clearCache merchant.id merchantOperatingCity.id serviceName
   logTagInfo "dashboard -> smsServiceConfigUpdate : " (show merchant.id)
