+++ Backend/app/rider-platform/rider-app/Main/src/Domain/Action/UI/Payment.hs
@@ -24,7 +24,7 @@ where
 import Control.Applicative ((<|>))
 import qualified Domain.Action.UI.FRFSTicketService as FRFSTicketService
 import qualified Domain.Types.Merchant as DM
-import qualified Domain.Types.Merchant.MerchantServiceConfig as DMSC
+import qualified Domain.Types.MerchantServiceConfig as DMSC
 import qualified Domain.Types.Person as DP
 import qualified Domain.Types.Ride as DRide
 import Environment
