+++ Backend/app/rider-platform/rider-app/Main/src/Storage/CachedQueries/Merchant/MerchantServiceConfig.hs
@@ -24,13 +24,23 @@ where
 import Data.Coerce (coerce)
 import Domain.Types.Common
 import Domain.Types.Merchant (Merchant)
-import Domain.Types.Merchant.MerchantServiceConfig
 import qualified Domain.Types.MerchantOperatingCity as DMOC
+import Domain.Types.MerchantServiceConfig
+import qualified Kernel.External.AadhaarVerification as AadhaarVerification
+import qualified Kernel.External.Call as Call
+import qualified Kernel.External.Maps.Interface.Types as Maps
+import qualified Kernel.External.Maps.Types as Maps
+import qualified Kernel.External.Notification as Notification
+import Kernel.External.Notification.Interface.Types as Notification
+import qualified Kernel.External.Payment.Interface as Payment
+import qualified Kernel.External.SMS.Interface as Sms
+import Kernel.External.Ticket.Interface.Types as Ticket
+import qualified Kernel.External.Whatsapp.Interface as Whatsapp
 import Kernel.Prelude
 import qualified Kernel.Storage.Hedis as Hedis
 import Kernel.Types.Id
 import Kernel.Utils.Common
-import qualified Storage.Queries.Merchant.MerchantServiceConfig as Queries
+import qualified Storage.Queries.MerchantServiceConfig as Queries
 
 -- create :: (MonadFlow m, CacheFlow m r, EsqDBFlow m r) => MerchantServiceConfig -> m ()
 -- create = Queries.create
@@ -61,6 +71,34 @@ cacheMerchantServiceConfig merchantServiceConfig = do
   expTime <- fromIntegral <$> asks (.cacheConfig.configsExpTime)
   let idKey = makeMerchantIdAndServiceKey merchantServiceConfig.merchantId merchantServiceConfig.merchantOperatingCityId (getServiceName merchantServiceConfig)
   Hedis.setExp idKey (coerce @MerchantServiceConfig @(MerchantServiceConfigD 'Unsafe) merchantServiceConfig) expTime
+  where
+    getServiceName :: MerchantServiceConfig -> ServiceName
+    getServiceName msc = case msc.serviceConfig of
+      MapsServiceConfig mapsCfg -> case mapsCfg of
+        Maps.GoogleConfig _ -> MapsService Maps.Google
+        Maps.OSRMConfig _ -> MapsService Maps.OSRM
+        Maps.MMIConfig _ -> MapsService Maps.MMI
+        Maps.NextBillionConfig _ -> MapsService Maps.NextBillion
+      SmsServiceConfig smsCfg -> case smsCfg of
+        Sms.ExotelSmsConfig _ -> SmsService Sms.ExotelSms
+        Sms.MyValueFirstConfig _ -> SmsService Sms.MyValueFirst
+        Sms.GupShupConfig _ -> SmsService Sms.GupShup
+      WhatsappServiceConfig whatsappCfg -> case whatsappCfg of
+        Whatsapp.GupShupConfig _ -> WhatsappService Whatsapp.GupShup
+      AadhaarVerificationServiceConfig aadhaarVerifictaionCfg -> case aadhaarVerifictaionCfg of
+        AadhaarVerification.GridlineConfig _ -> AadhaarVerificationService AadhaarVerification.Gridline
+      CallServiceConfig callCfg -> case callCfg of
+        Call.ExotelConfig _ -> CallService Call.Exotel
+      NotificationServiceConfig notificationCfg -> case notificationCfg of
+        Notification.FCMConfig _ -> NotificationService Notification.FCM
+        Notification.PayTMConfig _ -> NotificationService Notification.PayTM
+        Notification.GRPCConfig _ -> NotificationService Notification.GRPC
+      PaymentServiceConfig paymentCfg -> case paymentCfg of
+        Payment.JuspayConfig _ -> PaymentService Payment.Juspay
+      MetroPaymentServiceConfig paymentCfg -> case paymentCfg of
+        Payment.JuspayConfig _ -> MetroPaymentService Payment.Juspay
+      IssueTicketServiceConfig ticketCfg -> case ticketCfg of
+        Ticket.KaptureConfig _ -> IssueTicketService Ticket.Kapture
 
 makeMerchantIdAndServiceKey :: Id Merchant -> Id DMOC.MerchantOperatingCity -> ServiceName -> Text
 makeMerchantIdAndServiceKey id mocId serviceName = "CachedQueries:MerchantServiceConfig:MerchantId-" <> id.getId <> ":MechantOperatingCityId:-" <> mocId.getId <> ":ServiceName-" <> show serviceName
