+++ Backend/app/rider-platform/rider-app/Main/src/Tools/Call.hs
@@ -18,8 +18,8 @@ module Tools.Call
 where
 
 import Domain.Types.Merchant
-import qualified Domain.Types.Merchant.MerchantServiceConfig as DMSC
 import Domain.Types.MerchantOperatingCity
+import qualified Domain.Types.MerchantServiceConfig as DMSC
 import Domain.Types.MerchantServiceUsageConfig (MerchantServiceUsageConfig)
 import Kernel.External.Call as Reexport hiding
   ( initiateCall,
