+++ Backend/app/rider-platform/rider-app/Main/src/Tools/Maps.hs
@@ -28,8 +28,8 @@ module Tools.Maps
 where
 
 import Domain.Types.Merchant
-import qualified Domain.Types.Merchant.MerchantServiceConfig as DMSC
 import Domain.Types.MerchantOperatingCity (MerchantOperatingCity (..))
+import qualified Domain.Types.MerchantServiceConfig as DMSC
 import Domain.Types.MerchantServiceUsageConfig (MerchantServiceUsageConfig)
 import Domain.Types.Person (Person)
 import Kernel.External.Maps as Reexport hiding
