+++ Backend/app/rider-platform/rider-app/Main/src/Tools/SMS.hs
@@ -19,8 +19,8 @@ module Tools.SMS
 where
 
 import Domain.Types.Merchant
-import qualified Domain.Types.Merchant.MerchantServiceConfig as DMSC
 import qualified Domain.Types.MerchantOperatingCity as DMOC
+import qualified Domain.Types.MerchantServiceConfig as DMSC
 import Kernel.External.SMS as Reexport hiding
   ( sendSMS,
   )
