+++ Backend/app/rider-platform/rider-app/Main/src/Tools/Whatsapp.hs
@@ -21,8 +21,8 @@ module Tools.Whatsapp
 where
 
 import Domain.Types.Merchant
-import qualified Domain.Types.Merchant.MerchantServiceConfig as DMSC
 import qualified Domain.Types.MerchantOperatingCity as DMOC
+import qualified Domain.Types.MerchantServiceConfig as DMSC
 import Kernel.External.Types (ServiceFlow)
 import Kernel.External.Whatsapp.Interface as Reexport hiding
   ( whatsAppOptApi,
